a bunch of stuff

+++ assets/kernels/catmull.cu
@@ -1,4 +1,4 @@
-using Real = float;
+using Real = double;
 
 namespace catmull {
